packages/architect/jwt-provider/requirements.txt
pyjwt
flask
Werkzeug
JWCrypto
Flask-HTTPAuth
